content/base/test/test_bug364092.xhtml

<?xml version="1.0"?>
<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=364092
-->
<head>
  <title>Test for Bug 364092</title>
  <!-- XHTML needs explicit script elements -->
  <script type="text/javascript" src="/MochiKit/Base.js"></script>
  <script type="text/javascript" src="/MochiKit/Iter.js"></script>
  <script type="text/javascript" src="/MochiKit/DOM.js"></script>
  <script type="text/javascript" src="/MochiKit/Style.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=364092">Mozilla Bug 364092</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<div id="test1" foo="foo"/>
<pre id="test">
<script class="testbody" type="text/javascript">
/** Test for Bug 364092 **/

  var test1 = document.getElementById("test1");
  var attrNode = test1.getAttributeNode("foo");
  function mutationHandler(aEvent) {
    ok(attrNode == aEvent.relatedNode);
    ok(!test1.hasAttribute("foo"));
  }

  function runTest() {
    test1.addEventListener("DOMAttrModified", mutationHandler, true);
    test1.removeAttributeNode(attrNode);
    test1.removeEventListener("DOMAttrModified", mutationHandler, true);
  }

   runTest();
</script>
</pre>

</body>
</html>
